microservices/guests/requirements/common.txt
fastapi==0.105.0
httpx==0.25.2
psycopg2-binary==2.9.9
pydantic==2.5.2
python-dotenv==1.0.0
python-jose[cryptography]==3.3.0
SQLAlchemy==2.0.23
uvicorn==0.24.0
